Refactor metric normalizer tests
Move the methods that generate test metrics and the code that
compares metric inputs and normalized outputs to the `test_util`
metrics module. Previously these test helpers were duplicated
across DataDog, StatsD and AppSignal's metric sinks.

Rename the `run_comparisons` method to `assert_normalize`, as it
asserts the results of running a normalizer's `.normalize` method.

Move the duplicated test implementations to the `test_util::metrics`
module, in a separate tests sub-module, and make them generic over
the normalizer. Use these test definitions in the DataDog, StatsD
and AppSignal's metric sink tests.
